Refactor for ProfessorCard and ProjectImageBox
