+++ src/components/ProfessorCard/professorCard.css
@@ -1,4 +1,3 @@
-
 .professorCardOutline{
     background: var(--secondaryGreen);
     border-radius: 25px;
