+++ src/pages/ProjectDetailsPage/ProjectDetailsPage.tsx
@@ -6,7 +6,7 @@ import { getProject } from "../../services/getProject";
 import InformationBox from "../../components/ProjectDetail/InformationBox";
 import "./projectDetailPage.css";
 import Trax_Ghana from "../../images/Trax_Ghana.png";
-import ProjectImageBox from "../../components/ProjectImageBox";
+import ProjectImageBox from "../../components/ProjectImageBox/ProjectImageBox";
 import Meta from "../../components/Meta";
 
 const ProjectDetailsPage: React.FC = () => {
